requirements.txt
# Deep Learning Framework
torch>=2.2.0
torchvision>=0.17.0
torchaudio>=2.2.0

# Audio Processing
librosa>=0.10.1
soundfile>=0.12.1
audioread>=3.0.1

# Machine Learning & Clustering
scikit-learn>=1.3.0
scipy>=1.11.0

# Data Processing (Python 3.12 compatible versions)
numpy>=1.26.0
pandas>=2.0.0

# Visualization
matplotlib>=3.8.0
seaborn>=0.13.0

# Dimensionality Reduction
umap-learn>=0.5.5

# NLP for Lyrics (optional)
transformers>=4.30.0

# Utilities
tqdm>=4.66.0
joblib>=1.3.0

# Jupyter for notebooks
jupyter>=1.0.0
ipykernel>=6.25.0

# File handling
h5py>=3.10.0